+++ test/state_machine_test.dart
@@ -1,4 +1,4 @@
-// Copyright (c) 2017, Anatoly Pulyaevskiy. All rights reserved. Use of this source code
+// Copyright (c) 2018, Anatoly Pulyaevskiy. All rights reserved. Use of this source code
 // is governed by a BSD-style license that can be found in the LICENSE file.
 
 import 'package:redux_machine/redux_machine.dart';